feat: changed hibernate dao level, created hibernate.cfg.xml and HibernateUtil. Added log, with log4j.xml.

+++ src/main/java/org/matveyvs/dto/DirectionalDto.java
@@ -4,6 +4,6 @@ import org.matveyvs.entity.DownholeData;
 
 import java.sql.Timestamp;
 
-public record DirectionalDto(Long id, Timestamp measureDate, Double measureDepth,
+public record DirectionalDto(Integer id, Timestamp measureDate, Double measureDepth,
                              Double inc, Double azCorr, Double azTrue, DownholeData downholeData) {
 }
