+++ src/main/java/org/matveyvs/dto/DownholeDataDto.java
@@ -3,6 +3,6 @@ package org.matveyvs.dto;
 import org.matveyvs.entity.WellData;
 
 
-public record DownholeDataDto(Long id, WellData wellData) {
+public record DownholeDataDto(Integer id, WellData wellData) {
 
 }
