+++ src/main/java/org/matveyvs/dto/SurfaceDataDto.java
@@ -2,6 +2,6 @@ package org.matveyvs.dto;
 
 import java.sql.Timestamp;
 
-public record SurfaceDataDto(Long id, Timestamp measureDate, Double measureDepth,
+public record SurfaceDataDto(Integer id, Timestamp measureDate, Double measureDepth,
                              Double holeDepth, Double tvDepth) {
 }
