+++ src/main/java/org/matveyvs/dto/WellDataDto.java
@@ -1,5 +1,5 @@
 package org.matveyvs.dto;
 
-public record WellDataDto(Long id, String companyName,
+public record WellDataDto(Integer id, String companyName,
                           String fieldName, String wellCluster, String well) {
 }
